refactor: move core app files to src/app and add DESIGN.md
- Relocate main.tsx and home.tsx to src/app to align with Next.js-inspired structure
- Update index.html entry point and internal imports
- Create DESIGN.md documenting the Vercel/Next.js design system
- Remove obsolete src/pages directory
