+++ src/app/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
 import "@/styles/globals.css";
-import Home from "@/pages/home";
+import Home from "@/app/home";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
